B-16613: Decompose immutable RMS config into two classes to improve sonar LCOM.
